compile --watch

+++ src/commands/compile.js
@@ -8,9 +8,28 @@ module.exports = function (parameters, done) {
     var find = require("find");
     var path = require("path");
 
+
     var config = require("../config");
     var compiler = require("../functions/compiler");
     var notifier = require("../functions/notifier");
+    var functions = require("../functions/generateSourcesFunctions");
+
+    var currentlyCompiling = false;
+
+    if (parameters.watch) {
+        var watch = require('node-watch');
+        watch(config.meteorDirectory, {
+            followSymLinks: true,
+            filter: function (fullpath) {
+                return fullpath !== undefined && fullpath.indexOf(".meteor") === -1;
+            }
+        }, function (filename) {
+            if (functions.endsWith(filename, ".ts")) {
+                console.log(filename, ' changed.');
+                compileMeteorFiles();
+            }
+        });
+    }
 
     // var supersonicTargetFile = config.supersonicDirectory + "/www/scripts/smallstack.js";
     // var meteorTargetFile = config.meteorDirectory + "/shared/lib/smallstack.js";
@@ -18,7 +37,7 @@ module.exports = function (parameters, done) {
 
     function shallBeCompiled(type) {
         var keys = _.keys(parameters);
-        if (keys === undefined || keys.length === 0)
+        if (keys === undefined || (keys.indexOf("meteor") === -1 && keys.indexOf("smallstack") === -1 && keys.indexOf("supersonic") === -1))
             return true;
         return keys.indexOf(type) !== -1;
     }
@@ -47,14 +66,17 @@ module.exports = function (parameters, done) {
             console.log("compiling supersonic");
             compiler.compileTypescriptFiles(config.supersonicDirectory + "/www", { consolePrefix: "[supersonic]" }, nextFn);
         }
-        else
-            nextFn();
+        else {
+            if (typeof nextFn === 'function')
+                nextFn();
+        }
     }
 
     function compileMeteorFiles(nextFn) {
 
         // meteor files
-        if (shallBeCompiled("meteor") && config.meteorProjectAvailable()) {
+        if (shallBeCompiled("meteor") && config.meteorProjectAvailable() && !currentlyCompiling) {
+            currentlyCompiling = true;
             var meteorBuiltPath = path.join(config.tmpDirectory, "meteorbuilt");
             fs.removeSync(meteorBuiltPath);
             console.log("compiling meteor");
@@ -80,18 +102,27 @@ module.exports = function (parameters, done) {
                 if (fs.existsSync(path.join(meteorBuiltPath, "shared")))
                     fs.copySync(path.join(meteorBuiltPath, "shared"), path.join(config.meteorDirectory, "built", "shared"), { clobber: true });
 
-                nextFn();
+                if (parameters.watch)
+                    console.log("Watching : ", config.meteorDirectory);
+                else
+                    console.log("Meteor compilation complete!");
+
+                currentlyCompiling = false;
+
+                if (typeof nextFn === 'function')
+                    nextFn();
             });
         }
-        else
-            nextFn();
+        else {
+            if (typeof nextFn === 'function')
+                nextFn();
+        }
 
     }
 
     // compileSmallstackDataLayer(function () {
     //     compileSuperSonicFiles(function () {
     compileMeteorFiles(function () {
-        notifier("Compilation completed!");
         done();
     });
     //     });
